[MOVE] point 관련 파일들 package 이동
issue: #82
